+++ DatabaseCore/Endpoints/Posts/InsertShot.cs
@@ -6,20 +6,20 @@ namespace DatabaseCore.DatabaseComponents;
 
 public partial class RevMetrixDB
 {
-    public async Task<bool> InsertShot(int user_id,
-                                   int? frame_id,
-                                   int? ball_id,
-                                   int? video_id,
-                                   int? pins_remaining, 
+    public async Task<bool> InsertShot(long user_id,
+                                   long? frame_id,
+                                   long? ball_id,
+                                   long? video_id,
+                                   long? pins_remaining, 
                                    DateTime time,
-                                   int? lane_number,
-                                   float ddx,
-                                   float ddy,
-                                   float ddz,
-                                   float x_position,
-                                   float y_position,
-                                   float z_position,
-                                   int? pocket_hit
+                                   long? lane_number,
+                                   double ddx,
+                                   double ddy,
+                                   double ddz,
+                                   double x_position,
+                                   double y_position,
+                                   double z_position,
+                                   long? pocket_hit
         )
     {
         ConnectionString = Environment.GetEnvironmentVariable("SERVERDB_CONNECTION_STRING");
